# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "后台运营管理 API 文档",
    "description": "后台运营管理 Restful API",
    "version": "V1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:8008/api",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/system/role/modify": {
      "post": {
        "tags": [
          "角色服务",
          "修改角色"
        ],
        "operationId": "modifyRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/system/role/list": {
      "post": {
        "tags": [
          "角色服务",
          "角色列表"
        ],
        "operationId": "listRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleListReqVo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListRolePO"
                }
              }
            }
          }
        }
      }
    },
    "/system/role/grantRolePermission": {
      "post": {
        "tags": [
          "角色服务",
          "给角色分配权限"
        ],
        "operationId": "grantRolePermission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantPermissionReqVo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/system/role/created": {
      "post": {
        "tags": [
          "添加角色",
          "角色服务"
        ],
        "operationId": "createdRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/redis/setRedisValue": {
      "post": {
        "tags": [
          "设置redis值",
          "Redis服务"
        ],
        "operationId": "setRedisValue",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "value",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/redis/getRedisValue": {
      "post": {
        "tags": [
          "Redis服务",
          "获取redis值"
        ],
        "operationId": "getRedisValue",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GetRedisValueResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/role/remove": {
      "get": {
        "tags": [
          "角色服务",
          "删除角色"
        ],
        "operationId": "removeRole",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/system/role/getPermissionByRoleId": {
      "get": {
        "tags": [
          "根据角色获权限",
          "角色服务"
        ],
        "operationId": "getPermissionByRoleId",
        "parameters": [
          {
            "name": "roleId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseListPermissionPo"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BaseResponse": {
        "type": "object",
        "properties": {
          "接口状态码": {
            "type": "integer",
            "format": "int32"
          },
          "提示信息": {
            "type": "string"
          },
          "数据": {
            "type": "object"
          },
          "总计": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "RolePO": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "okDel": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roleName": {
            "type": "string"
          },
          "roleRemark": {
            "type": "string"
          },
          "hasRole": {
            "type": "boolean"
          },
          "permission": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
      "RoleListReqVo": {
        "type": "object",
        "properties": {
          "current": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "roleName": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "BaseResponseListRolePO": {
        "type": "object",
        "properties": {
          "接口状态码": {
            "type": "integer",
            "format": "int32"
          },
          "提示信息": {
            "type": "string"
          },
          "数据": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RolePO"
            }
          },
          "总计": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "GrantPermissionReqVo": {
        "type": "object",
        "properties": {
          "角色id": {
            "type": "integer",
            "format": "int64"
          },
          "权限id": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "GetRedisValueResponse": {
        "type": "object",
        "properties": {
          "接口状态码": {
            "type": "integer",
            "format": "int32"
          },
          "提示信息": {
            "type": "string"
          },
          "数据": {
            "type": "object"
          },
          "总计": {
            "type": "integer",
            "format": "int64"
          },
          "value": {
            "type": "string"
          }
        }
      },
...
      "BaseResponseListPermissionPo": {
        "type": "object",
        "properties": {
          "接口状态码": {
            "type": "integer",
            "format": "int32"
          },
          "提示信息": {
            "type": "string"
          },
          "数据": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionPo"
            }
          },
          "总计": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PermissionPo": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "okDel": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "permissionType": {
            "type": "integer",
            "format": "int32"
          },
          "permissionCode": {
            "type": "string"
          },
          "permissionName": {
            "type": "string"
          },
          "resourceUrl": {
            "type": "string"
          },
          "menuIcon": {
            "type": "string"
          },
          "menuSortnum": {
            "type": "integer",
            "format": "int32"
          },
          "permissionRemark": {
            "type": "string"
          },
          "hasPermission": {
            "type": "boolean"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionPo"
            }
          }
        }
      }
    }
  }
}